% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Returns the attributes of the concept in character format}
\usage{
getAttributes(concepts, idx)
}
\arguments{
\item{concepts}{Concept Lattice by fcaR}

\item{idx}{number of a concept}
}
\value{
attributes of the concept
}
\description{
Returns the attributes of the concept in character format
}
\examples{
library(fcaR)
fc <- FormalContext$new(vegas)
fc$find_concepts()
concepts <- fc$concepts
getAttributes(concepts,59)
}
