% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdx.R
\name{getIdx}
\alias{getIdx}
\title{Returns the index corresponding to the given set of attributes}
\usage{
getIdx(concepts, atr)
}
\arguments{
\item{concepts}{Concept Lattice by fcaR}

\item{atr}{Set of attributes}
}
\value{
index of the Concept Lattice
}
\description{
Returns the index corresponding to the given set of attributes
}
\examples{
library(fcaR)
fc <- FormalContext$new(vegas)
fc$find_concepts()
concepts <- fc$concepts
atr <- concepts$sub(122)
getIdx(concepts, atr$get_intent())
}
