% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_sublattice.R
\name{graph_sublattice}
\alias{graph_sublattice}
\title{Returns a graph given a sublattice from fcaR}
\usage{
graph_sublattice(sublattice)
}
\arguments{
\item{sublattice}{sublattice from package fcaR}
}
\value{
graph type igraph
}
\description{
Returns a graph given a sublattice from fcaR
}
\examples{
library(fcaR)
library(igraph)
fc <- FormalContext$new(vegas)
fc$find_concepts()
concepts <- fc$concepts
idx <- which(fc$concepts$support() > 0.9)
sublattice <- fc$concepts$sublattice(idx)
graph_sublattice(sublattice)
}
