% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Shiny Application}
\usage{
launchApp()
}
\value{
This function does not return anything. It initiates a Shiny app
}
\description{
Shiny Application
}
\examples{
\dontrun{
launchApp()
}
}
