% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_by_attributes.R
\name{recommend_by_attributes}
\alias{recommend_by_attributes}
\title{Recommendation Based on the Best Confidence Match for One or Multiple Attributes}
\usage{
recommend_by_attributes(fc, selectedAttributes, probAttributes)
}
\arguments{
\item{fc}{a FormalContext using fcaR}

\item{selectedAttributes}{the desired attributes}

\item{probAttributes}{the attributes to check}
}
\value{
a data frame with the best possible recommendation for each attribute to check
}
\description{
Recommendation Based on the Best Confidence Match for One or Multiple Attributes
}
\examples{
library(fcaR)
fc <- FormalContext$new(vegas)
recommend_by_attributes(fc, c("Traveler type=Friends","Gym"), c("Score=5"))
recommend_by_attributes(fc,
                        c("Traveler type=Friends","Pool", "Spa"),
                        c("Score=3", "Score=4", "Score=5"))
}
