% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_by_max_cardinality.R
\name{recommend_by_max_cardinality}
\alias{recommend_by_max_cardinality}
\title{Recommendation Based on Maximum Cardinality with Minimum Confidence}
\usage{
recommend_by_max_cardinality(fc, selectedAttributes, conf_value)
}
\arguments{
\item{fc}{a FormalContext using fcaR}

\item{selectedAttributes}{the desired attributes}

\item{conf_value}{a value between 0 and 1 to filter concepts by that confidence}
}
\value{
a data frame with the best recommendations sorted by the number of attributes the concept possesses
}
\description{
Recommendation Based on Maximum Cardinality with Minimum Confidence
}
\examples{
library(fcaR)
fc <- FormalContext$new(vegas)
recommend_by_max_cardinality(fc, c("Traveler type=Friends","Gym"), 0.5)
recommend_by_max_cardinality(fc, c("Pool", "Spa", "Score=4"), 0.6)
}
